src/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
exceptiongroup==1.2.2
faiss-cpu==1.10.0
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.10
jinja2==3.1.6
joblib==1.4.2
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pygments==2.19.1
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.51.3
typer==0.15.2
typing-extensions==4.13.2
typing-inspection==0.4.0
urllib3==2.4.0
uvicorn==0.34.2
